fix: rename leftover dashkoor to data-control-center
