+++ api/resources/timestamp/timestamp.proto
@@ -4,7 +4,7 @@
 syntax = "proto3";
 
 package resources.timestamp;
-option go_package = "github.com/koor-tech/data-control-center/gen/go/proto/resources/timestamp;timestamp";
+option go_package = "github.com/koor-tech/data-control-center/gen/go/api/resources/timestamp;timestamp";
 
 import "google/protobuf/timestamp.proto";
