+++ api/services/auth/auth.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 package services.auth;
-option go_package = "github.com/koor-tech/dashcore/gen/go/proto/services/auth;auth";
+option go_package = "github.com/koor-tech/data-control-center/gen/go/api/services/auth;auth";
 
 import "api/resources/timestamp/timestamp.proto";
 import "validate/validate.proto";
