+++ api/services/stats/stats.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 package services.cluster;
-option go_package = "github.com/koor-tech/dashcore/gen/go/proto/services/cluster;cluster";
+option go_package = "github.com/koor-tech/data-control-center/gen/go/api/services/cluster;cluster";
 
 message StreamRequest {}
